requirements.txt
asgiref==3.8.1
Django==5.1.2
django-environ==0.11.2
djangorestframework==3.15.2
mysqlclient==2.2.5
sqlparse==0.5.1
